requirements.txt
asgiref==3.8.1
borb==2.1.22
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
crayons==0.4.0
cryptography==42.0.6
Django==5.0
django-rosetta==0.10.0
django-parler==2.3
django-jazzmin==3.0.0
engineering-notation==0.10.0
et-xmlfile==1.1.0
fonttools==4.51.0
gunicorn==22.0.0
idna==3.7
lxml==5.2.1
mysql-connector-python==8.3.0
mysqlclient==2.2.4
mysqlx-connector-python==8.3.0
numpy==1.26.4
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==24.0
pandas==2.1.2
piexif==1.1.3
pillow==10.3.0
protobuf==4.21.12
pyasn1==0.6.0
pycparser==2.22
pycryptodome==3.20.0
pydub==0.25.1
pypng==0.20220715.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
pytz==2024.1
qrcode==7.4.2
requests==2.31.0
rsa==4.9
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
stegano==0.11.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
xlrd==2.0.1
XlsxWriter==3.1.9
xlutils==2.0.0
xlwt==1.3.0
xlsxdocument==1.3.10